--- src/index.js
@@ -1,2 +1,0 @@
-import React from 'react'
-export {ConfigProvider, useConfig} from "./useConfig"
