+++ src/index.ts
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export { ConfigProvider, useConfig } from './useConfig';
